Change method arg type

+++ app/src/main/java/com/ict/mito/gootravel/spot/select/radar/ui/dialog/SelectSpotBottomSheetViewModel.kt
@@ -16,6 +16,6 @@ class SelectSpotBottomSheetViewModel(private val repository: Repository) : ViewM
 
     val distance: String = "0m"
 
-    fun setId(id: Int) {
+    fun setId(id: Long) {
     }
 }
